Changed return type of SList::append from bool to void.
Ideally, append should return a bool if it fails. This will require checking if, for example, the construction of the node to be appended fails. For now I'm going for the simpler, less robust approach by returning void. Interestingly, g++ did not catch the lack of a return value in this code.